Organização de views , por views de renderização e views de retorno json
